backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ren <84360463+rencryptofish@users.noreply.github.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
gitpython = "^3.1.31"
psycopg2 = "^2.9.6"
peewee = "^3.16.0"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
isort = "^5.12.0"
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
